+++ bin/package.js
@@ -81,12 +81,12 @@ const all = {
   asar: {
     // A glob expression, that unpacks the files with matching names to the
     // "app.asar.unpacked" directory.
-    unpack: 'WebTorrent*'
+    unpack: 'NFTTorrent*'
   },
 
   // The build version of the application. Maps to the FileVersion metadata property on
   // Windows, and CFBundleVersion on Mac. Note: Windows requires the build version to
-  // start with a number. We're using the version of the underlying WebTorrent library.
+  // start with a number. We're using the version of the underlying NFTTorrent library.
   buildVersion: require('webtorrent/package.json').version,
 
   // The application source directory.
@@ -121,14 +121,14 @@ const darwin = {
   arch: 'x64',
 
   // The bundle identifier to use in the application's plist (Mac only).
-  appBundleId: 'io.webtorrent.webtorrent',
+  appBundleId: 'xyz.graviton.webtorrent',
 
   // The application category type, as shown in the Finder via "View" -> "Arrange by
   // Application Category" when viewing the Applications directory (Mac only).
   appCategoryType: 'public.app-category.utilities',
 
   // The bundle identifier to use in the application helper's plist (Mac only).
-  helperBundleId: 'io.webtorrent.webtorrent-helper',
+  helperBundleId: 'xyz.graviton.webtorrent-helper',
 
   // Application icon.
   icon: config.APP_ICON + '.icns'
@@ -223,6 +223,12 @@ function buildDarwin (cb) {
         CFBundleURLIconFile: path.basename(config.APP_FILE_ICON) + '.icns',
         CFBundleURLName: 'BitTorrent Stream-Magnet URL',
         CFBundleURLSchemes: ['stream-magnet']
+      },
+      {
+        CFBundleTypeRole: 'Editor',
+        CFBundleURLIconFile: path.basename(config.APP_FILE_ICON) + '.icns',
+        CFBundleURLName: 'NFTTorrent URL',
+        CFBundleURLSchemes: ['nft-torrent']
       }
     ]
 
@@ -267,6 +273,7 @@ function buildDarwin (cb) {
     function signApp (cb) {
       const sign = require('electron-osx-sign')
       const { notarize } = require('electron-notarize')
+      const { appleId, appleIdPassword, teamId } = require('./secrets');
 
       /*
        * Sign the app with Apple Developer ID certificates. We sign the app for 2 reasons:
@@ -285,18 +292,21 @@ function buildDarwin (cb) {
         verbose: true,
         app: appPath,
         platform: 'darwin',
-        identity: 'Developer ID Application: WebTorrent, LLC (5MAMC8G3L8)',
+        identity: 'Developer ID Application: Graviton Inc. (V75C78WU28)',
         hardenedRuntime: true,
         entitlements: path.join(config.ROOT_PATH, 'bin', 'darwin-entitlements.plist'),
         'entitlements-inherit': path.join(config.ROOT_PATH, 'bin', 'darwin-entitlements.plist'),
-        'signature-flags': 'library'
+        'signature-flags': 'library',
+        'gatekeeper-assess': false
       }
 
       const notarizeOpts = {
         appBundleId: darwin.appBundleId,
         appPath,
-        appleId: 'feross@feross.org',
-        appleIdPassword: '@keychain:AC_PASSWORD'
+        appleId,
+        appleIdPassword,
+        teamId,
+        tool: 'notarytool'
       }
 
       console.log('Mac: Signing app...')
@@ -538,13 +548,13 @@ function buildLinux (cb) {
       src: filesPath + '/',
       dest: DIST_PATH,
       arch: destArch,
-      bin: 'WebTorrent',
+      bin: 'NFTTorrent',
       icon: {
         '48x48': path.join(config.STATIC_PATH, 'linux/share/icons/hicolor/48x48/apps/webtorrent-desktop.png'),
         '256x256': path.join(config.STATIC_PATH, 'linux/share/icons/hicolor/256x256/apps/webtorrent-desktop.png')
       },
       categories: ['Network', 'FileTransfer', 'P2P'],
-      mimeType: ['application/x-bittorrent', 'x-scheme-handler/magnet', 'x-scheme-handler/stream-magnet'],
+      mimeType: ['application/x-bittorrent', 'x-scheme-handler/magnet', 'x-scheme-handler/stream-magnet', 'x-scheme-handler/nft-torrent'],
       desktopTemplate: path.join(config.STATIC_PATH, 'linux/webtorrent-desktop.ejs'),
       lintianOverrides: [
         'unstripped-binary-or-object',
@@ -582,13 +592,13 @@ function buildLinux (cb) {
       src: filesPath + '/',
       dest: DIST_PATH,
       arch: destArch,
-      bin: 'WebTorrent',
+      bin: 'NFTTorrent',
       icon: {
         '48x48': path.join(config.STATIC_PATH, 'linux/share/icons/hicolor/48x48/apps/webtorrent-desktop.png'),
         '256x256': path.join(config.STATIC_PATH, 'linux/share/icons/hicolor/256x256/apps/webtorrent-desktop.png')
       },
       categories: ['Network', 'FileTransfer', 'P2P'],
-      mimeType: ['application/x-bittorrent', 'x-scheme-handler/magnet', 'x-scheme-handler/stream-magnet'],
+      mimeType: ['application/x-bittorrent', 'x-scheme-handler/magnet', 'x-scheme-handler/stream-magnet', 'x-scheme-handler/nft-torrent'],
       desktopTemplate: path.join(config.STATIC_PATH, 'linux/webtorrent-desktop.ejs')
     }