+++ bin/secrets.example.js
@@ -1,0 +1,5 @@
+module.exports = {
+  "appleId": "braulio@graviton.xyz", // Email of the Apple ID used to sign the app
+  "appleIdPassword": "123456",  // App-specific password or keychain reference to a App-specific password for the Apple ID used to sign the app
+  "teamId": "ABC123" // Apple team id
+}