+++ src/renderer/pages/preferences-page.js
@@ -3,12 +3,16 @@ const PropTypes = require('prop-types')
 
 const colors = require('material-ui/styles/colors')
 const Checkbox = require('material-ui/Checkbox').default
-const RaisedButton = require('material-ui/RaisedButton').default
 const Heading = require('../components/heading')
+const Wallet = require('../components/wallet')
 const PathSelector = require('../components/path-selector')
+const CustomButton = require('../components/custom-button')
 
 const { dispatch } = require('../lib/dispatcher')
 const config = require('../../config')
+
+const defaultCheckboxIconStyle = {fill: 'rgba(17, 13, 33, 0.2)'};
+const defaultCheckedIcon = (<img src='chk_icon.svg' alt='chk' width={24} height={24} />);
 
 class PreferencesPage extends React.Component {
   constructor (props) {
@@ -55,8 +59,10 @@ class PreferencesPage extends React.Component {
       <Preference>
         <Checkbox
           className='control'
+          iconStyle={defaultCheckboxIconStyle}
+          checkedIcon={defaultCheckedIcon}
           checked={!this.props.state.saved.prefs.openExternalPlayer}
-          label='Play torrent media files using WebTorrent'
+          label='Play torrent media files using NFT Torrent'
           onCheck={this.handleOpenExternalPlayerChange}
         />
       </Preference>
@@ -72,6 +78,8 @@ class PreferencesPage extends React.Component {
       <Preference>
         <Checkbox
           className='control'
+          iconStyle={defaultCheckboxIconStyle}
+          checkedIcon={defaultCheckedIcon}
           checked={this.props.state.saved.prefs.highestPlaybackPriority}
           label='Highest Playback Priority'
           onCheck={this.handleHighestPlaybackPriorityChange}
@@ -118,6 +126,8 @@ class PreferencesPage extends React.Component {
       <Preference>
         <Checkbox
           className='control'
+          iconStyle={defaultCheckboxIconStyle}
+          checkedIcon={defaultCheckedIcon}
           checked={this.props.state.saved.prefs.autoAddTorrents}
           label='Watch for new .torrent files and add them immediately'
           onCheck={(e, value) => { this.handleAutoAddTorrentsChange(e, value) }}
@@ -171,17 +181,17 @@ class PreferencesPage extends React.Component {
     if (isFileHandler) {
       return (
         <Preference>
-          <p>WebTorrent is your default torrent app. Hooray!</p>
+          <p>NFT Torrent is your default torrent app. Hooray!</p>
         </Preference>
       )
     }
     return (
       <Preference>
-        <p>WebTorrent is not currently the default torrent app.</p>
-        <RaisedButton
+        <p>NFT Torrent is not currently the default torrent app.</p>
+        <CustomButton
           className='control'
           onClick={this.handleSetDefaultApp}
-          label='Make WebTorrent the default'
+          label='Make NFT Torrent the default'
         />
       </Preference>
     )
@@ -189,6 +199,14 @@ class PreferencesPage extends React.Component {
 
   handleStartupChange (e, isChecked) {
     dispatch('updatePreferences', 'startup', isChecked)
+  }
+  setWallet() {
+    const account = this.props.state?.saved?.auth?.address;
+    return (
+      <Preference>
+        <Wallet state={this.props.state} account={account}/>
+      </Preference>
+    );
   }
 
   setStartupCheckbox () {
@@ -200,8 +218,10 @@ class PreferencesPage extends React.Component {
       <Preference>
         <Checkbox
           className='control'
+          iconStyle={defaultCheckboxIconStyle}
+          checkedIcon={defaultCheckedIcon}
           checked={this.props.state.saved.prefs.startup}
-          label='Open WebTorrent on startup'
+          label='Open NFT Torrent on startup'
           onCheck={this.handleStartupChange}
         />
       </Preference>
@@ -213,6 +233,8 @@ class PreferencesPage extends React.Component {
       <Preference>
         <Checkbox
           className='control'
+          iconStyle={defaultCheckboxIconStyle}
+          checkedIcon={defaultCheckedIcon}
           checked={this.props.state.saved.prefs.soundNotifications}
           label='Enable sounds'
           onCheck={this.handleSoundNotificationsChange}
@@ -231,12 +253,15 @@ class PreferencesPage extends React.Component {
 
   render () {
     const style = {
-      color: colors.grey400,
-      marginLeft: 25,
-      marginRight: 25
+      color: 'rgba(255, 255, 255, 0.5)',
+      marginLeft: 16,
+      marginRight: 16,
     }
     return (
       <div style={style}>
+        <PreferencesSection title='Wallet'>
+          {this.setWallet()}
+        </PreferencesSection>
         <PreferencesSection title='Folders'>
           {this.downloadPathSelector()}
           {this.autoAddTorrentsCheckbox()}
@@ -268,12 +293,18 @@ class PreferencesSection extends React.Component {
 
   render () {
     const style = {
-      marginBottom: 25,
-      marginTop: 25
+      // marginBottom: 25,
+      marginTop: 10
+    }
+    const headingStyle = {
+      padding: 12,
+      borderBottom: '1px solid rgba(255, 255, 255, 0.1)',
+      margin: 0,
+      fontSize: 16,
     }
     return (
-      <div style={style}>
-        <Heading level={2}>{this.props.title}</Heading>
+      <div style={style} className='box'>
+        <Heading level={2} style={headingStyle}>{this.props.title}</Heading>
         {this.props.children}
       </div>
     )
@@ -282,7 +313,7 @@ class PreferencesSection extends React.Component {
 
 class Preference extends React.Component {
   render () {
-    const style = { marginBottom: 10 }
+    const style = { margin: 12 }
     return (<div style={style}>{this.props.children}</div>)
   }
 }